+++ src/main/java/racingcar/application/RacingCarPositions.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.application;
 
 import java.util.ArrayList;
 import java.util.List;
