+++ src/main/java/racingcar/application/RacingMain.java
@@ -1,4 +1,7 @@
-package racingcar;
+package racingcar.application;
+
+import racingcar.view.InputView;
+import racingcar.view.ResultView;
 
 public class RacingMain {
     public static void main(String[] args){
