+++ src/main/java/racingcar/utils/Const.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.utils;
 
 public class Const {
     public static final int INITIAL_NUM = 0;
