+++ src/main/java/racingcar/view/ResultView.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.view;
+
+import racingcar.application.RacingCarPositions;
 
 import java.util.List;
