+++ src/test/java/racingcar/RacingCarTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.domain.RacingCar;
 
 import static org.assertj.core.api.Assertions.assertThat;
