+++ src/test/java/racingcar/RandomUtilsTest.java
@@ -2,6 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.utils.RandomUtils;
 
 import static org.assertj.core.api.Assertions.assertThat;
